backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
aiohttp==3.9.1
prometheus-client==0.19.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
pydantic==2.5.0
email-validator==2.1.0
python-dateutil==2.8.2
google-generativeai==0.3.1